prediction_server/requirements.txt
matplotlib==3.0.3
scipy==1.2.1
numpy==1.16.2
scikit-learn[alldeps]
pandas==0.25.0
websockets==8.0.2
Flask==1.1.1
requests==2.22.0
redis==3.3.5
rq==1.1.0
cassandra-driver==3.18
cqlengine==0.21.0
imbalanced-learn==0.5.0
pyzmq==18.1.0
pika==1.1.0